Move termToBatch in Analyzer Fix variable name
