+++ app/analyzers/terms.py
@@ -4,7 +4,7 @@ from helpers.singletons import settings, es, logging
 from collections import defaultdict
 from collections import Counter
 import helpers.utils
-from helpers.analyzer import Analyzer
+from helpers.analyzer import Analyzer, add_term_to_batch
 
 
 class TermsAnalyzer(Analyzer):
@@ -59,9 +59,8 @@ class TermsAnalyzer(Analyzer):
 
                     for aggregator_sentence in aggregator_sentences:
                         flattened_aggregator_sentence = helpers.utils.flatten_sentence(aggregator_sentence)
-                        eval_terms_array = helpers.utils.add_term_to_batch(eval_terms_array,
-                                                                           flattened_aggregator_sentence,
-                                                                           flattened_target_sentence, observations, doc)
+                        eval_terms_array = add_term_to_batch(eval_terms_array, flattened_aggregator_sentence,
+                                                             flattened_target_sentence, observations, doc)
 
                 total_terms_added += len(target_sentences)
