+++ app/helpers/analyzer.py
@@ -1,5 +1,8 @@
 import abc
+from collections import defaultdict
 from configparser import NoOptionError
+from typing import DefaultDict, Optional, Dict
+
 from helpers.singletons import settings, es, logging
 import helpers.utils
 from helpers.outlier import Outlier
@@ -111,3 +114,15 @@ class Analyzer(abc.ABC):
     @abc.abstractmethod
     def evaluate_model(self):
         raise NotImplementedError()
+
+
+def add_term_to_batch(eval_terms_array: DefaultDict, aggregator_value: Optional[str], target_value: Optional[str],
+                      observations: Dict, doc: Dict) -> DefaultDict:
+    if aggregator_value not in eval_terms_array.keys():
+        eval_terms_array[aggregator_value] = defaultdict(list)
+
+    eval_terms_array[aggregator_value]["targets"].append(target_value)
+    eval_terms_array[aggregator_value]["observations"].append(observations)
+    eval_terms_array[aggregator_value]["raw_docs"].append(doc)
+
+    return eval_terms_array
