+++ app/helpers/utils.py
@@ -5,11 +5,7 @@ import numpy as np
 import base64
 import re
 from statistics import mean, median
-import os
 import validators
-from collections import defaultdict
-
-from typing import DefaultDict, Dict, Optional
 
 
 def flatten_dict(d, parent_key='', sep='.'):
@@ -290,15 +286,3 @@ def nested_dict_values(d):
             yield from nested_dict_values(v)
         else:
             yield v
-
-
-def add_term_to_batch(eval_terms_array: DefaultDict, aggregator_value: Optional[str], target_value: Optional[str],
-                      observations: Dict, doc: Dict) -> DefaultDict:
-    if aggregator_value not in eval_terms_array.keys():
-        eval_terms_array[aggregator_value] = defaultdict(list)
-
-    eval_terms_array[aggregator_value]["targets"].append(target_value)
-    eval_terms_array[aggregator_value]["observations"].append(observations)
-    eval_terms_array[aggregator_value]["raw_docs"].append(doc)
-
-    return eval_terms_array
